Fix -Wsign-compare warnings
Add casts at the signed/unsigned comparisons that "-Wsign-compare"
flags.  Change the strbin2hex() and strhex2bin() loop counters from int
to size_t.  Change unmunge's "t_len" the same way.  Each now matches
the value it is compared against.  "t_len" is assigned from strftime(),
which returns size_t, so the int was already a narrowing conversion.

The assertions in "replay.c" need more than a cast.  They compare the
cred's "mac_len" against a sizeof with ">=".  A negative "mac_len"
would convert to a large unsigned value and pass, so test for a
non-negative "mac_len" first.

Assert a non-negative length in _hkdf_expand() and dec_validate_mac()
as well.  Those comparisons already fail on a negative length, so
the new assertions catch nothing new.  They document an invariant
that another file establishes.

+++ src/common/mac_test.c
@@ -69,7 +69,7 @@ check_mac (munge_mac_t m, const char *str, const char *key, int keylen,
         fail ("check_mac invalid dst or dstlen for %s", str);
         return -1;
     }
-    if (dstlen > sizeof buf) {
+    if ((size_t) dstlen > sizeof buf) {
         fail ("check_mac %ld-byte buf too small for %s %d-byte result",
                 sizeof buf, str, dstlen);
         return -1;
